auto variable

+++ cocos2dx/Classes/HelloWorldScene.h
@@ -86,7 +86,7 @@ public:
     
     float randomValueBetween(float low, float high);
     void setInvisible(cocos2d::Node * node);
-    float getTimeTick();
+    double getTimeTick();
     
     virtual void setTouchEnabled(bool enable);
     virtual void onTouchesBegan(const std::vector<cocos2d::Touch*>& touches, cocos2d::Event* event);
